Semantic changes
Aimed to address #6. Changes header from "Log In" to "Who are you?". In
addition provides a tagline on both the login and register pages to
explain the reason that identify details are needed. Also clarifies
alerts to suggest visiting the "Who are you?" page when a user does not
have permission to perform an action.